refactor(vendo): wire route table — grants/approvals (kill-list B4)
